Clippy-Gate gruen + Zeroize-Haertung Basis (CryptographicInformation, hmac_sign)
- signature!/verify! Closures zu Bloecken (kein ? im Body), ?-Makros unberuehrt
- type_complexity via HybridTestBytes-Alias behoben
- Drop+zeroize fuer CryptographicInformation (content, passphrase)
- Drop+zeroize fuer SignatureData und SignatureKey

+++ src/legacy/kyber_crypto/kyber_crypto_aes.rs
@@ -34,9 +34,9 @@ where
         }
 
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -69,9 +69,9 @@ where
         passphrase: &str,
     ) -> Result<(Vec<u8>, Vec<u8>), CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -103,9 +103,9 @@ where
         passphrase: &str,
     ) -> Result<(Vec<u8>, Vec<u8>), CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -200,9 +200,9 @@ where
         }
 
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
@@ -235,9 +235,9 @@ where
         ciphertext: Vec<u8>,
     ) -> Result<Vec<u8>, CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
@@ -269,9 +269,9 @@ where
         ciphertext: Vec<u8>,
     ) -> Result<Vec<u8>, CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
