+++ src/legacy/kyber_crypto/kyber_crypto_aes_xts.rs
@@ -35,9 +35,9 @@ where
         }
 
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -70,9 +70,9 @@ where
         passphrase: &str,
     ) -> Result<(Vec<u8>, Vec<u8>), CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -104,9 +104,9 @@ where
         passphrase: &str,
     ) -> Result<(Vec<u8>, Vec<u8>), CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -202,9 +202,9 @@ where
         }
 
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
@@ -237,9 +237,9 @@ where
         ciphertext: Vec<u8>,
     ) -> Result<Vec<u8>, CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
@@ -271,9 +271,9 @@ where
         ciphertext: Vec<u8>,
     ) -> Result<Vec<u8>, CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
