+++ src/legacy/kyber_crypto/kyber_crypto_xchacha_poly.rs
@@ -35,9 +35,9 @@ where
         }
 
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -72,9 +72,9 @@ where
         passphrase: &str,
     ) -> Result<(Vec<u8>, Vec<u8>), CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -108,9 +108,9 @@ where
         passphrase: &str,
     ) -> Result<(Vec<u8>, Vec<u8>), CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::encryption(),
@@ -209,9 +209,9 @@ where
         }
 
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
@@ -245,9 +245,9 @@ where
         ciphertext: Vec<u8>,
     ) -> Result<Vec<u8>, CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
@@ -281,9 +281,9 @@ where
         ciphertext: Vec<u8>,
     ) -> Result<Vec<u8>, CryptError> {
         let (key_encap_mechanism, kybersize) = match KyberSize::variant() {
-            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512 as usize),
-            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768 as usize),
-            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024 as usize),
+            KyberVariant::Kyber512 => (KeyEncapMechanism::kyber512(), 512_usize),
+            KyberVariant::Kyber768 => (KeyEncapMechanism::kyber768(), 768_usize),
+            KyberVariant::Kyber1024 => (KeyEncapMechanism::kyber1024(), 1024_usize),
         };
         let crypt_metadata = CryptographicMetadata::from(
             Process::decryption(),
